+++ src/main/java/com/github/jaksonlin/testcraft/domain/annotations/NullValue.java
@@ -1,0 +1,10 @@
+package com.github.jaksonlin.testcraft.domain.annotations;
+
+import com.alibaba.fastjson.annotation.JSONType;
+
+@JSONType(typeName = "NullValue")
+public class NullValue extends DefaultValue {
+    public NullValue() {
+        super("NullValue");
+    }
+}