+++ CHANGELOG.md
@@ -74,3 +74,8 @@
 
 ## 1.0.10 - 2025-05-06
 - bugfix, git command execution in a separate thread to avoid UI blocking
+
+## 1.0.11 - 2025-05-13
+- lib dependency upgrade, use fastjson to parse the annotation schema
+- put all toolwindow into one multi-tab toolwindow
+- invalid test case scan on directory